requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy
scikit-learn
pyttsx3
pillow